rename #7: component 파일 이름 정리
