+++ android/app/src/main/java/com/example/passedpath/feature/place/presentation/screen/AddPlaceScreen.kt
@@ -41,7 +41,7 @@ import com.example.passedpath.feature.place.presentation.state.AddPlaceUiState
 import com.example.passedpath.feature.place.presentation.viewmodel.AddPlaceViewModel
 import com.example.passedpath.feature.place.presentation.viewmodel.AddPlaceViewModelFactory
 import com.example.passedpath.ui.component.button.BaseButton
-import com.example.passedpath.feature.place.presentation.component.PlaceSearchCard
+import com.example.passedpath.feature.place.presentation.component.PlaceSearchResultCard
 import com.example.passedpath.feature.place.presentation.component.PlaceSearchTextField
 import com.example.passedpath.ui.theme.Gray100
 import com.example.passedpath.ui.theme.Gray500
@@ -217,7 +217,7 @@ private fun SearchResultList(
                 items = uiState.places,
                 key = PlaceSearchResult::stableKey
             ) { place ->
-                PlaceSearchCard(
+                PlaceSearchResultCard(
                     title = place.name,
                     address = place.displayAddress,
                     isSelected = place.stableKey == uiState.selectedPlaceId,
