+++ android/app/src/main/java/com/example/passedpath/feature/route/presentation/component/RouteEmptyToast.kt
@@ -1,4 +1,4 @@
-package com.example.passedpath.ui.component.toast
+package com.example.passedpath.feature.route.presentation.component
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
@@ -14,6 +14,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.passedpath.R
+import com.example.passedpath.ui.component.toast.BaseToast
 import com.example.passedpath.ui.theme.PassedPathTheme
 
 @Composable
